Added area component, updated old area component to change name to addarea component
