+++ src/area.js
@@ -1,22 +1,25 @@
 import React from 'react';
 // import {Link} from 'react-router-dom';
 
-class AddArea extends React.Component {
+class AreaPage extends React.Component {
     render() {
         return (
-            <div className='addArea'>
-                <h3>Add Area</h3>
-                <p>Enter Name of Area</p>
-                <form>
-                    <label>
-                        Area:
-                        <input type="text" name="areaName" />
-                    </label>
-                    <input type="submit" value="Confirm" />
-                </form>
+            <div className='area'>
+                <div className='areaHeader'>
+                    <h2>Area Name</h2>
+                    <div className='sectCount'>
+                        <p>Num</p>
+                    </div>
+                    <div className='placementCount'>
+                        <p>Num</p>
+                    </div>
+                </div>
+                <div className='areaBody'>
+                <input type="button" className="btn" value="Add Section"/>
+                </div>
             </div>
         );
     }
 }
 
-export default AddArea;
+export default AreaPage;
